Changed rotation adjustment buttons to 45 and 22.5, redesigned layout
Changed the rotation adjustment buttons from 30 and 45 degrees to 22.5 and 45 degrees to fit with player head rotation states in Minecraft. Also changed internals so that all rotation adjustments go through 1 function with an argument. This will allow for more flexible angle additions later if necessary. Also changed the layout of the rotation adjustment buttons to look less squished. Not sure if I like it though.